fix(win): guard POSIX fork/pipe/poll in persistent executor; fix size_t/ul mismatch
  - persistent_process_executor.cpp: wrap poll.h, sys/wait.h includes and
    all fork/execvp/waitpid/kill/usleep function bodies with #ifndef _WIN32;
    start() throws 'not supported on Windows' to propagate cleanly
  - error_reporter.cpp:218: std::max(1ul, ...) fails on MinGW64 where size_t
    is unsigned long long but 1ul is unsigned long; changed to size_t(1)

+++ src/semantic/error_reporter.cpp
@@ -215,7 +215,7 @@ std::string ErrorReporter::formatWithContext(const Diagnostic& diag) const {
 
     // Source code context
     if (diag.show_source && !source_lines_.empty()) {
-        size_t start_line = std::max(1ul, diag.line - diag.context_lines);
+        size_t start_line = std::max(size_t(1), diag.line - diag.context_lines);
         size_t end_line = std::min(source_lines_.size(), diag.line + diag.context_lines);
 
         // Line number width for alignment
